refactor : infrastructure 패키지 생성 및 리팩토링
